requirements.txt
pillow
requests
numpy<2.0
opencv-python